implemented formatError
