+++ test/test.ts
@@ -46,7 +46,6 @@ describe('Login Mutation test', async () => {
     const password = 'dumb_password';
     const hashedPassword = hash(password, email + process.env.XSALT);
     const query = requestLogin(email, password, false);
-    await credentialValidation(email, hashedPassword);
 
     const response = await request(url).post('').send(query);
 
@@ -58,16 +57,6 @@ describe('Login Mutation test', async () => {
     expect(response.body.data.login.user.cpf).to.be.eq('1');
   });
 });
-
-async function credentialValidation(email: string, hashedPassword: string) {
-  if (!validateEmail(email)) {
-    throw new Error('Invalid email format.');
-  }
-  const userWithSameCredentials = await getRepository(User).findOne({ email, password: hashedPassword });
-  if (!userWithSameCredentials) {
-    throw new Error('Wrong credentials.');
-  }
-}
 
 function requestLogin(email: string, password: string, rememberMe: boolean) {
   return {
